MXBackgroundSyncService: Create MXSyncResponseStoreManager to manage store logic
